ingress/requirements.txt
# networking service
uvloop==0.16.0
msgpack==1.0.3
cloudpickle==2.0.0
pyzmq==22.3.0
aiozmq==0.9.0
aiojobs==1.0.0
# monitoring service
psutil==5.9.0